% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_paths.R
\name{absolute_path}
\alias{absolute_path}
\title{Function to return absolute paths from relative paths.}
\usage{
absolute_path(files)
}
\arguments{
\item{files}{A vector of file paths.}
}
\description{
Returns absolute paths from relative paths.
}
