% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ARG_codes.R
\name{classify_ARG_codes}
\alias{classify_ARG_codes}
\title{classify_ARG_codes}
\usage{
classify_ARG_codes(codes)
}
\arguments{
\item{codes}{vector or dataframe column of ARG codes.}
}
\description{
This function checks if the rows in a dataframe are contained in the other.
Generally most useful for testing if the outputs of 2 functions are =.
}
