% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ARG_codes.R
\name{classify_ARG_genes}
\alias{classify_ARG_genes}
\title{classify_ARG_genes}
\usage{
classify_ARG_genes(genes, obo = NULL)
}
\arguments{
\item{genes}{vector or dataframe column of ARG ARO accession numbers or gene names.}

\item{obo}{an object processed with \code{\link{processOBO}}.}
}
\description{
Classifies ARGs from the \href{https://card.mcmaster.ca/home}{CARD database} using either accession number or gene name.
}
