% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.r
\name{dna_comp}
\alias{dna_comp}
\title{Compliment}
\usage{
dna_comp(DNA)
}
\arguments{
\item{DNA}{A string or vector of strings of DNA}
}
\value{
phyloseq-object
}
\description{
produces the compliment sequence of a DNA strand
}
