% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{plot_sequence_quality}
\alias{plot_sequence_quality}
\title{Plot Sequence Quality}
\usage{
sequence_quality(file)
}
\arguments{
\item{file}{a list created by sequence_quality()}
}
\value{
ggplot-object
}
\description{
Creates plots of sequence qualitires
}
