% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quality_report.R
\name{read_quality_report}
\alias{read_quality_report}
\title{read_quality_report}
\usage{
read_quality_report(path, q, n, k)
}
\arguments{
\item{path}{File path(s) to fastq or fastq.gz file(s).}

\item{q}{Quality score cutoff for the read. Will look at the mean average score for 
\code{k} bases beyond where the potential read length cutoff would be recommended.}

\item{k}{number of bases beyond the current to consider for quality cutoff.}

\item{n}{The number of records to sample from the fastq file.}

\item{cores}{The number of CPU cores/threads to use.}
}
\value{
data.table
}
\description{
Processes fastq files to look at read counts, read lengths,
and at what read cycle the quality drops below the \code{q}
quality threshold.
}
