% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_NA.R
\name{replace_NA}
\alias{replace_NA}
\title{replace_NA}
\usage{
replace_NA(DT, replace_with = 0)
}
\arguments{
\item{DT}{A data.table.}

\item{replace_with}{value to replace NA with.}
}
\description{
Function to replace NA values in numeric columns of a data.table
}
