% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.r
\name{rev_comp}
\alias{rev_comp}
\title{Reverse Compliment}
\usage{
rev_comp(DNA)
}
\arguments{
\item{DNA}{A string or vector of strings of DNA}
}
\value{
phyloseq-object
}
\description{
produces the reverse compliment sequence of a DNA strand
}
