% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_match.R
\name{row_match}
\alias{row_match}
\title{row_match}
\usage{
row_match(dataframe_1, dataframe_2, results = "in")
}
\arguments{
\item{dataframe_1}{A dataframe.}

\item{dataframe_2}{A dataframe..}

\item{results}{what the function should return; 'in', 'not_in', 'num_in', 'num_not_in'}
}
\description{
This function checks if the rows in a dataframe are contained in the other.
Generally most useful for testing if the outputs of 2 functions are =.
}
\keyword{row_match}
