% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schuylR-package.R
\docType{package}
\name{schuylR}
\alias{schuylR}
\alias{schuylR-package}
\title{schuylR-package}
\description{
schuylR is a conglomeration of functions written to do
various things in R that I find useful to use frequently.
}
\details{
\url{https://schuyler-smith.github.io/schuylR/}
}
\author{
Schuyler Smith \email{schuyler.smith@iastate.edu}
}
\keyword{package}
