% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{sequence_quality}
\alias{sequence_quality}
\title{Sequence Quality}
\usage{
sequence_quality(file, n = 5e+05)
}
\arguments{
\item{file}{A string with either a file name or directory containing the 
required fastq file types.}

\item{n}{The number of reads to sample when processing FASTQ files.}
}
\value{
list
}
\description{
Reads in ".fastq.gz", ".fastq.bz2", or ".fastq" files to create a list of
dataframes for analysis of sequence qualities
}
