% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unload_all_packages.R
\name{unload_all_packages}
\alias{unload_all_packages}
\title{unload_all_packages}
\usage{
unload_all_packages()
}
\description{
This function unloads all library'd packages from your environment.
}
\keyword{packages}
\keyword{unload}
