% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{Function for creating color palettes for graphs.}
\usage{
create_palette(color_count, colors = 'default')
}
\arguments{
\item{color_count}{Number of colors to choose for palette.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palete of R-accepted
colors.}
}
\value{
palette
}
\description{
Creates color palettes.
}
